Refactor: Move button components to ui/buttons directory structure
- Moved all button-3d variants to buttons/3d directory
- Moved specialized button components to their own directories (colorful, shimmer, rainbow)
- Created detailed README files for each button category
- Added backward compatibility with @deprecated notices
- Updated COMPONENT-MOVEMENT-PLAN.md to track progress
