+++ src/components/ui/buttons/3d/README.md
@@ -1,0 +1,21 @@
+# 3D Button Components
+
+This directory contains various 3D button components with different depth effects.
+
+## Components
+
+- `button-3d.tsx` - Basic 3D button with depth effect
+- `button-3d-border.tsx` - 3D button with distinct borders
+- `button-3d-enhanced.tsx` - 3D button with sophisticated depth effects
+- `button-3d-physical.tsx` - 3D button with physical, realistic appearance
+
+## Usage
+
+```tsx
+import {
+  Button3D,
+  Button3DBorder,
+  Button3DEnhanced,
+  Button3DPhysical
+} from '@/components/ui/buttons/3d';
+```
