+++ src/components/ui/buttons/3d/index.ts
@@ -1,0 +1,4 @@
+export { Button3D } from './button-3d';
+export { Button3DBorder } from './button-3d-border';
+export { Button3DEnhanced } from './button-3d-enhanced';
+export { Button3DPhysical } from './button-3d-physical';
