+++ src/components/ui/buttons/colorful/README.md
@@ -1,0 +1,13 @@
+# Colorful Button Components
+
+This directory contains button components with colorful gradient effects.
+
+## Components
+
+- `button-colorful.tsx` - Button with gradient background effect
+
+## Usage
+
+```tsx
+import { ButtonColorful } from '@/components/ui/buttons/colorful';
+```
