+++ src/components/ui/buttons/colorful/index.ts
@@ -1,0 +1,1 @@
+export { ButtonColorful } from './button-colorful';
