+++ src/components/ui/buttons/index.ts
@@ -1,1 +1,5 @@
-export * from './button'
+export * from './button';
+export * from './3d';
+export * from './colorful';
+export * from './shimmer';
+export * from './rainbow';
