+++ src/components/ui/buttons/rainbow/README.md
@@ -1,0 +1,13 @@
+# Rainbow Button Components
+
+This directory contains button components with rainbow animation effects.
+
+## Components
+
+- `rainbow-button.tsx` - Button with rainbow gradient animation
+
+## Usage
+
+```tsx
+import { RainbowButton } from '@/components/ui/buttons/rainbow';
+```
