+++ src/components/ui/buttons/rainbow/index.ts
@@ -1,0 +1,1 @@
+export { RainbowButton } from './rainbow-button';
