+++ src/components/ui/buttons/shimmer/README.md
@@ -1,0 +1,13 @@
+# Shimmer Button Components
+
+This directory contains button components with shimmer/glow effects.
+
+## Components
+
+- `shimmer-button.tsx` - Button with shimmer animation on hover
+
+## Usage
+
+```tsx
+import { ShimmerButton } from '@/components/ui/buttons/shimmer';
+```
