+++ src/components/ui/buttons/shimmer/index.ts
@@ -1,0 +1,1 @@
+export { ShimmerButton } from './shimmer-button';
